Continued work game calc
Made the method evaluateKick public. Got the parameters set up to send
from the GameCard into the GameController. Want to add more wind speed
directions, i.e. southwest, northwest, southeast, north east, but left
as is for simplicity right now. added a button that we can use to submit
the game information and run through the evaluteKick and evaluateGoal.
Need to figure out the win box dimmensions in a way that is easy to
remember and scale. I'm working on it, started messing around with the
graphics again this evening.